requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-socketio
werkzeug
pyjwt
python-dotenv
